Put todo files into a todo folder
