--- src/components/css/text-edit.css
@@ -1,12 +1,0 @@
-.text-edit {
-    white-space: pre-wrap;
-    word-break: break-all;
-    word-wrap: break-word;
-    overflow-wrap: break-word;
-}
-
-.text-edit:empty:before {
-    content: attr(placeholder);
-    color: #9E9E9E;
-    cursor: text;
-}
