--- src/components/css/todo-components.css
@@ -1,168 +1,0 @@
-/** ********** ***** ********** **/
-/** ********** BOARD ********** **/
-/** ********** ***** ********** **/
-
-/** ********** Container & Generic children rules ********** **/
-
-.todo-board {
-    display: flex;
-    flex-direction: column;
-    padding: .5rem;
-    max-width: 100%;
-}
-
-.todo-board > * {
-    max-width: 100%;
-}
-
-/** ********** Top Control ********** **/
-
-.todo-board-top-control {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-}
-
-.todo-board-top-control > * {
-    max-width: 100%;
-}
-
-/** ********** Logs ********** **/
-
-#todo-board-logs {
-    width: 100%;
-    height: 5rem;
-    padding: .25rem;
-    background-color: #212121;
-    overflow: scroll;
-}
-
-.todo-board-log {
-    font: .75em Consolas, monospace;
-    color: #fafafa;
-    text-align: left;
-}
-
-/** ********** Title ********** **/
-
-.todo-board-title {
-    margin: .5rem;
-    text-align: center;
-}
-
-/** ********** Create-New ********** **/
-
-.todo-create-new {
-    display: flex;
-    flex-direction: column;
-
-    margin: .5rem;
-    width: 25rem;
-    border: 1px solid #757575;
-
-    box-shadow: 0 4px 8px #BDBDBD;
-}
-
-/** ********** Cards ********** **/
-
-.todo-cards {
-    display: flex;
-    align-items: flex-start;
-    justify-content: flex-start;
-    flex-flow: row wrap;
-
-    margin: .5rem;
-}
-
-/** ********** ********** ********** **/
-/** ********** TO-DO CARD ********** **/
-/** ********** ********** ********** **/
-
-.todo-card {
-    position: relative;
-
-    display: flex;
-    flex-direction: column;
-
-    margin: .5rem;
-    width: 25rem; max-height: 50rem;
-    border: 1px solid #757575;
-
-    box-shadow: 0 4px 8px #BDBDBD;
-
-    transition: box-shadow .25s, border-color .25s;
-}
-
-.todo-card:focus {
-    outline: none;
-}
-
-.todo-card:hover {
-    box-shadow: 0 4px 12px #757575;
-}
-
-.todo-card-title {
-    padding: .5rem;
-    font-size: 1.1rem;
-    text-align: left;
-}
-
-.todo-card-list-items {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    text-align: left;
-}
-
-.todo-card-list-item {
-    display: flex;
-    align-items: center;
-
-    margin: .5rem 0;
-}
-
-.todo-card-list-item > * {
-    margin: 0 .5rem;
-}
-
-.todo-card-list-item-checkbox {
-    display: flex;
-    align-items: center;
-}
-
-.todo-card-list-item-description {
-    flex-grow: 1;
-
-    white-space: pre-wrap;
-    word-break: break-all;
-    word-wrap: break-word;
-    overflow-wrap: break-word;
-
-    cursor: default;
-}
-
-/** ********** ***************** ********** **/
-/** ********** TO-DO EDIT WINDOW ********** **/
-/** ********** ***************** ********** **/
-
-/** ********** Container & Generic children rules ********** **/
-
-.todo-edit-window {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-/** ********** Backdrop ********** **/
-
-.todo-edit-window-backdrop {
-    position: fixed;
-    top: 0;
-    left: 0;
-
-    width: 100vw;
-    height: 100vh;
-
-    background-color: rgba(224, 224, 224, .75);
-}
