+++ src/components/todo/css/options-panel.css
@@ -1,0 +1,96 @@
+/** ********** ************* ********** **/
+/** ********** OPTIONS PANEL ********** **/
+/** ********** ************* ********** **/
+
+/** ********** Container & Generic children rules ********** **/
+
+.todo-options-panel {
+  position: relative;
+  display: flex;
+  justify-content: flex-start;
+  padding: 0.5rem;
+}
+
+/** ********** Options panel buttons ********** **/
+
+.todo-options-panel-btn {
+  position: relative;
+  padding: 0 0.25rem;
+  font-size: 1.5rem;
+  cursor: pointer;
+  user-select: none;
+}
+
+.todo-options-panel-btn:hover .todo-options-panel-btn-tooltip {
+  visibility: visible;
+  opacity: 1;
+}
+
+.todo-options-panel-btn:focus {
+  outline: none;
+}
+
+/** ********** Options panel button tooltips ********** **/
+
+.todo-options-panel-btn-tooltip {
+  visibility: hidden;
+
+  position: absolute;
+  left: 50%;
+  transform: translate(-50%, 25%);
+
+  border-radius: 5px;
+  padding: 0.25rem;
+
+  background-color: #212121;
+
+  font-size: 0.6rem;
+  color: #fafafa;
+  text-align: center;
+
+  white-space: nowrap;
+
+  opacity: 0;
+  transition: opacity 0.25s;
+}
+
+/** ********** Options panel color palette ********** **/
+
+.todo-options-panel-color-palette {
+  position: absolute;
+  left: 50%;
+  transform: translate(-50%, -135%);
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-flow: row wrap;
+
+  width: 5rem;
+  border: 1px solid #bdbdbd;
+
+  background-color: #fafafa;
+  border-radius: 5px;
+
+  transition: opacity 0.25s;
+}
+
+.todo-options-panel-color-palette.hide {
+  visibility: hidden;
+  opacity: 0;
+}
+
+.todo-options-panel-color-palette-btn {
+  margin: 0.25rem;
+  width: 1rem;
+  height: 1rem;
+  border: 1px solid #fafafa;
+
+  border-radius: 50%;
+
+  transition: border-color 0.25s;
+}
+
+.todo-options-panel-color-palette-btn:hover {
+  border-color: #bdbdbd;
+}
