+++ src/components/todo/css/todo-edit.css
@@ -5,39 +5,39 @@
 /** ********** Container & Generic children rules ********** **/
 
 .todo-edit {
-    display: flex;
-    flex-direction: column;
+  display: flex;
+  flex-direction: column;
 
-    margin: .5rem;
-    width: 25rem;
-    border: 1px solid #757575;
+  margin: 0.5rem;
+  width: 25rem;
+  border: 1px solid #757575;
 
-    box-shadow: 0 4px 8px #757575;
+  box-shadow: 0 4px 8px #757575;
 }
 
 .todo-edit.hidden {
-    display: none;
+  display: none;
 }
 
 .todo-edit.hidden-vis {
-    visibility: hidden;
+  visibility: hidden;
 }
 
 /** ********** Title ********** **/
 
 .todo-edit-title {
-    padding: .5rem;
-    font-size: 1.1rem;
-    text-align: left;
+  padding: 0.5rem;
+  font-size: 1.1rem;
+  text-align: left;
 }
 
 /** ********** List of items ********** **/
 
 .todo-edit-list-items {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    text-align: left;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  text-align: left;
 }
 
 /** ********** ******************** ********** **/
@@ -47,58 +47,59 @@
 /** ********** Container & Generic children rules ********** **/
 
 .todo-edit-list-item {
-    display: flex;
-    align-items: center;
+  display: flex;
+  align-items: center;
 
-    margin: .5rem 0;
+  margin: 0.5rem 0;
 }
 
 .todo-edit-list-item:focus-within {
-    outline: 1px solid #9E9E9E;
+  outline: 1px solid #9e9e9e;
 }
 
 .todo-edit-list-item > * {
-    margin: 0 .5rem;
+  margin: 0 0.5rem;
 }
 
 /** ********** Dragger & Deleter ********** **/
 
-.todo-edit-list-item-dragger, .todo-edit-list-item-deleter {
-    visibility: hidden;
+.todo-edit-list-item-dragger,
+.todo-edit-list-item-deleter {
+  visibility: hidden;
 }
 
 .todo-edit-list-item-dragger {
-    cursor: grab;
+  cursor: grab;
 }
 
 .todo-edit-list-item-deleter {
-    cursor: pointer;
+  cursor: pointer;
 }
 
 .todo-edit-list-item:hover .todo-edit-list-item-dragger,
 .todo-edit-list-item:hover .todo-edit-list-item-deleter,
 .todo-edit-list-item:focus-within .todo-edit-list-item-deleter {
-    visibility: visible;
+  visibility: visible;
 }
 
 /** ********** Checkbox ********** **/
 
 .todo-edit-list-item-checkbox {
-    display: flex;
-    align-items: center;
+  display: flex;
+  align-items: center;
 }
 
 /** ********** Description (text) ********** **/
 
 .todo-edit-list-item-description {
-    flex-grow: 1;
+  flex-grow: 1;
 
-    white-space: pre-wrap;
-    word-break: break-all;
-    word-wrap: break-word;
-    overflow-wrap: break-word;
+  white-space: pre-wrap;
+  word-break: break-all;
+  word-wrap: break-word;
+  overflow-wrap: break-word;
 }
 
 .todo-edit-list-item-description:focus {
-    outline: none;
+  outline: none;
 }
